Restructure based on review comments
